Update const.py
Restructured, with children generally falling in the 1-10 range, PG in 10-13, teens in 13-17, and adults as 17+

Potential issues:
"A" has conflicts. All viewers in Spain. In other regions, it's adult.
"R" is equivalent to PG17 in the US, but is adult in Canada/Taiwan.
"C" is recommended for children in Thailand, but PG-15 in Japan (video games only)